Fix cmd-build-bitcode defined commands.
I found that this architecture is not good.
For example, jni/Application.mk cannot use ARM and LLVM at the
same time. This bug is because of the rules in build/core/.

I will do a big refactoring on <GDK>/build/core and de-coupling
this from <NDK>/build/core these days.