Update samples
The motivation of this update was to remove libnativehelper which was made
implicitly available recently.  But since we are on it, update samples
source to recent one which allows same bitcode to compile/run on
ARM/X86/MIPS.  Tiny arch-specific *so are still required to call libbcc.
Future llvm ABI in ndk will remove this restriction and phase out this.

Change-Id: Ic32d7dcfbbe421c07911590b22dfd865cdeb2fb0